big modif page tache

+++ database/migrations/2026_05_01_120035_create_calendriers_table.php
@@ -6,6 +6,7 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration {
     public function up(): void {
+        if (Schema::hasTable('calendriers')) return;
         Schema::create('calendriers', function (Blueprint $table) {
             $table->id();
             $table->foreignId('user_id')->constrained()->onDelete('cascade');
