+++ database/migrations/2026_05_01_123858_add_user_id_to_taches_table.php
@@ -11,6 +11,7 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasColumn('taches', 'user_id')) return;
         Schema::table('taches', function (Blueprint $table) {
             $table->foreignId('user_id')->nullable()->constrained()->onDelete('cascade');
         });
